Adding handleexit.py to infra
